게시글 지역구분 탭 기능 추가

+++ backend/src/main/java/com/jibangyoung/domain/admin/dto/AdRegionDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AdRegionDTO {
-    private Integer region_code;
+    private Long region_code;
     private String sido; 
 }
